ocr_bus_number/requirements.txt
fastapi
uvicorn[standard]
onnxruntime
torch
transformers
numpy
opencv-python-headless
pillow
requests
python-dotenv
sentencepiece
python-multipart